% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covCSE.r
\name{lmvgamma}
\alias{lmvgamma}
\title{Log multivariate gamma function}
\usage{
lmvgamma(a, p)
}
\arguments{
\item{a}{a numeric scalar.}

\item{p}{a positive integer.}
}
\value{
a scalar
}
\description{
Compute the logarithm of the multivariate gamma function 
\eqn{\log \Gamma_p(a)}.
}
\author{
Peter Hoff
}
