% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{msqrt}
\alias{msqrt}
\title{Symmetric square root of a matrix}
\usage{
msqrt(M)
}
\arguments{
\item{M}{a positive semidefinite matrix.}
}
\value{
a positive semidefinite matrix.
}
\description{
Compute the symmetric square root of a matrix.
}
\examples{
S<-rWishart(1,5,diag(5))[,,1]
S 
Sh<-msqrt(S)
Sh\%*\%Sh 

}
\author{
Peter Hoff
}
