% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{msqrtInv}
\alias{msqrtInv}
\title{Inverse symmetric square root of a matrix}
\usage{
msqrtInv(M)
}
\arguments{
\item{M}{a positive definite matrix.}
}
\value{
a positive definite matrix.
}
\description{
Compute the inverse of the symmetric square root of a matrix.
}
\examples{
S<-rWishart(1,5,diag(5))[,,1]
solve(S) 
iSh<-msqrtInv(S)
iSh\%*\%iSh 

}
\author{
Peter Hoff
}
